packages/vector_server/pyproject.toml
[tool.poetry]
name = "VectorServer"
version = "0.1.0"
description = ""
authors = [
    "Felipe Novaes Rocha <felipe.novaes.rocha@gmail.com>",
]

[tool.poetry.dependencies]
python = ">=3.9.0,<=3.12"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
#google-api-python-client = "^2.88.0"
#faiss-cpu = "^1.7.4"
#sentence-transformers = "^2.2.2"
#jsonschema = "^4.17.3"
#base58 = "^2.1.1"
deeplake = "^3.6.1"
tiktoken = "^0.4.0"
fastapi = "^0.96.0"
langchain = "^0.0.193"
uvicorn = { extras = ["standard"], version = "^0.22.0" }
#marvin = "^0.9.2"
#python-jose = { extras = ["cryptography"], version = "^3.3.0" }
#siwe = "^2.2.0"
#web3 = "^6.4.0"
asgi-correlation-id = "^4.2.0"
#ray = { extras = ["serve"], version = "^2.5.1" }
pymongo = "^4.3.3"
backoff = "^2.2.1"
tc-messagebroker = "^1.5.1"


[tool.poetry.group.dev.dependencies]
flake8 = "6.0.0"
black = "23.3.0"
isort = "5.12.0"


[tool.pdm]
[tool.pdm.dev-dependencies]
dev = []

[tool.pdm.build]
includes = []
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
authors = [
    {name = "Felipe Novaes Rocha", email = "felipe.novaes.rocha@gmail.com"},
]
requires-python = ">=3.9,<3.12.1"
dependencies = [
    "python-dotenv<2.0.0,>=1.0.0",
    "deeplake<4.0.0,>=3.6.1",
    "fastapi<1.0.0,>=0.96.0",
    "uvicorn[standard]<1.0.0,>=0.22.0",
    "asgi-correlation-id<5.0.0,>=4.2.0",
    "pymongo<5.0.0,>=4.3.3",
    "backoff<3.0.0,>=2.2.1",
    "tc-messagebroker<2.0.0,>=1.5.1",
    "openai<1.0.0,>=0.27.8",
    "celery[redis]>=5.3.4",
    "tiktoken<1.0.0,>=0.4.0",
    "langchain<1.0.0,>=0.0.193",
]
name = "VectorServer"
version = "0.1.0"
description = ""

[project.group.dev.dependencies]
flake8 = "6.0.0"
black = "23.3.0"
isort = "5.12.0"
